Extract test values from command factories into TestValues.
